# Conflicts:
#	.idea/misc.xml
#	src/main/java/items/RangedWeapon.java

+++ src/main/java/items/Weapon.java
@@ -1,12 +1,10 @@
 package items;
 
-import items.Item;
-
 public abstract class Weapon extends Item {
 
     public Weapon(String name, String description) {
         super(name, description);
     }
 
-    public abstract double getDamage();
+    public abstract int getDamage();
 }
